need to work algo out

+++ Subject.java
@@ -12,6 +12,6 @@ public interface Subject
     boolean attach(Observer obs);
     boolean detach(Observer obs);
     boolean notifyAllObservers();
-    public void setData(int recDepth, int radiusRatio, Color cactusColor, Color pearColor);
+    public void setData(int recDepth, double radiusRatio, Color cactusColor, Color pearColor);
     ArrayList<FractalElement> getData();
 }
